refactored non cartesian training to merge utilities between unet and pdnet
