+++ fastmri_recon/tests/training_scripts/ncpdnet_train_test.py
@@ -1,7 +1,7 @@
 import pytest
 
-from fastmri_recon.training_scripts import ncpdnet_train
-from fastmri_recon.training_scripts.ncpdnet_train import train_ncpdnet
+from fastmri_recon.training_scripts import nc_train
+from fastmri_recon.training_scripts.nc_train import train_ncpdnet
 
 
 @pytest.mark.parametrize('kwargs',[
@@ -10,10 +10,10 @@ from fastmri_recon.training_scripts.ncpdnet_train import train_ncpdnet
     {'dcomp': True},
 ])
 def test_train_ncpdnet(create_full_fastmri_test_tmp_dataset, kwargs):
-    ncpdnet_train.FASTMRI_DATA_DIR = create_full_fastmri_test_tmp_dataset['fastmri_tmp_data_dir']
-    ncpdnet_train.LOGS_DIR = create_full_fastmri_test_tmp_dataset['logs_tmp_dir']
-    ncpdnet_train.CHECKPOINTS_DIR = create_full_fastmri_test_tmp_dataset['checkpoints_tmp_dir']
-    ncpdnet_train.n_volumes_train = 2
+    nc_train.FASTMRI_DATA_DIR = create_full_fastmri_test_tmp_dataset['fastmri_tmp_data_dir']
+    nc_train.LOGS_DIR = create_full_fastmri_test_tmp_dataset['logs_tmp_dir']
+    nc_train.CHECKPOINTS_DIR = create_full_fastmri_test_tmp_dataset['checkpoints_tmp_dir']
+    nc_train.n_volumes_train = 2
     train_ncpdnet(
         af=8,
         n_samples=2,
