+++ fastmri_recon/tests/training_scripts/unet_nc_train_test.py
@@ -1,18 +1,18 @@
 import pytest
 
-from fastmri_recon.training_scripts import unet_nc_train
-from fastmri_recon.training_scripts.unet_nc_train import train_unet
+from fastmri_recon.training_scripts import nc_train
+from fastmri_recon.training_scripts.nc_train import train_unet_nc
 
 
 @pytest.mark.parametrize('kwargs',[
     {'dcomp': True},
 ])
-def test_train_unet(create_full_fastmri_test_tmp_dataset, kwargs):
-    unet_nc_train.FASTMRI_DATA_DIR = create_full_fastmri_test_tmp_dataset['fastmri_tmp_data_dir']
-    unet_nc_train.LOGS_DIR = create_full_fastmri_test_tmp_dataset['logs_tmp_dir']
-    unet_nc_train.CHECKPOINTS_DIR = create_full_fastmri_test_tmp_dataset['checkpoints_tmp_dir']
-    unet_nc_train.n_volumes_train = 2
-    train_unet(
+def test_train_unet_nc(create_full_fastmri_test_tmp_dataset, kwargs):
+    nc_train.FASTMRI_DATA_DIR = create_full_fastmri_test_tmp_dataset['fastmri_tmp_data_dir']
+    nc_train.LOGS_DIR = create_full_fastmri_test_tmp_dataset['logs_tmp_dir']
+    nc_train.CHECKPOINTS_DIR = create_full_fastmri_test_tmp_dataset['checkpoints_tmp_dir']
+    nc_train.n_volumes_train = 2
+    train_unet_nc(
         af=8,
         n_samples=2,
         n_epochs=1,
